+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 1.23.4
 ### Features
 * Added a new `interactive-repl` value to the `container-type` label
 * Added `ForInteractiveREPLContainer` functions to the `EnclaveObjectLabelsProvider` and `EnclaveObjectNameProvider`